feat: find statment endpoint

+++ domain/entity/client.go
@@ -3,6 +3,6 @@ package entity
 type Client struct {
 	ID      int    `json:"id"`
 	Name    string `json:"nome"`
-	Limit   int64  `json:"limite"`
-	Balance int64  `json:"saldo"`
+	Limit   int    `json:"limite"`
+	Balance int    `json:"saldo"`
 }
